epedev-301 preview in RB
add checking for query param dialog to determine if we are in regular RB
mode or in iframe dialog mode.  also add a new variable to determine if
we should show tabs or not.  when dialog has a assigned tab value, we
hide the tabs otherwise we assume all tabs are require to show