3.2 llb add resource bug fix
this fix addresses LLB add dataset functionality issue.  the issue was
that when clicking on different add resource type button, the iframed RB
only display the result of the 1st resource type.

the solution is to separate out regular resource browser controller from
LLB or CM resource browser controller.  the code are duplicated with
some modification for LLB & CM usage.  not a great solution but this
does address the issue for now.

This also address the bug report epedev-405